PR-43 Moss comments addressed
