+++ linfa-ensemble/README.md
@@ -1,9 +1,5 @@
 # Ensemble learning
 
-`linfa-ensemble` provides methods for ensemble learning algorithms.
-
-## The Big Picture
-
 `linfa-ensemble` is a crate in the [`linfa`](https://crates.io/crates/linfa) ecosystem, an effort to create a toolkit for classical Machine Learning implemented in pure Rust, akin to Python's `scikit-learn`.
 
 ## Current state
@@ -19,6 +15,9 @@ There is an example in the `examples/` directory how to use random forest models
 ```bash
 $ cargo run --release --example random_forest
 ```
+
+The current benchmark performs random forest with 10, 100, 500, 1000 independent trees to predict labels and compare such results to ground truth.
+
 
 ## License
 Dual-licensed to be compatible with the Rust project.
