+++ linfa-ensemble/src/lib.rs
@@ -1,5 +1,5 @@
-mod random_forest;
-mod voting_classifier;
+pub mod random_forest;
+pub mod voting_classifier;
 
 pub use random_forest::*;
 pub use voting_classifier::*;
