+++ linfa-ensemble/src/random_forest/hyperparameters.rs
@@ -57,7 +57,6 @@ impl RandomForestParamsBuilder {
     }
 
     pub fn build(&self) -> RandomForestParams {
-        // TODO
         let max_features = self.max_features.unwrap_or(MaxFeatures::Auto);
 
         let max_features = match max_features {
